+++ CHANGELOG.md
@@ -1,5 +1,42 @@
 # CHANGELOG
 
+## v2.4.0 [WIP]
+
+**Improvements**
+- It's possible to retrieve the instances of _Attachment_, _Bookmark_, _Thumbnail_, _Toolbar_ plugins from the _Default Layout_ plugin instance.
+
+~~~ javascript
+const defaultLayoutPluginInstance = defaultLayoutPlugin();
+
+const {
+    attachmentPluginInstance,
+    bookmarkPluginInstance,
+    thumbnailPluginInstance,
+    toolbarPluginInstance,
+} = defaultLayoutPluginInstance;
+~~~
+
+Similarity, the _Toolbar_ plugin instance provides the accesses to the instance of other plugins that build the toolbar:
+
+~~~ javascript
+const toolbarPluginInstance = toolbarPlugin();
+
+const {
+    dropPluginInstance,
+    fullScreenPluginInstance,
+    getFilePluginInstance,
+    openPluginInstance,
+    pageNavigationPluginInstance,
+    printPluginInstance,
+    propertiesPluginInstance,
+    rotatePluginInstance,
+    scrollModePluginInstance,
+    searchPluginInstance,
+    selectionModePluginInstance,
+    zoomPluginInstance,
+} = toolbarPluginInstance;
+~~~
+
 ## v2.3.2
 
 **Improvements**